+++ mote/views.py
@@ -115,7 +115,10 @@ class ElementIframeView(ElementBaseView):
     """Element view suitable for rendering in an iframe"""
 
     def get_template_names(self):
-        return ["mote/element/iframe.html"]
+        return [
+            "%s/mote/element/iframe.html" % self.element.project.id,
+            "mote/element/iframe.html"
+        ]
 
 
 class VariationBaseView(ElementBaseView):