+++ mote/tests/mote/projects/demo/website/src/main.js
@@ -1,0 +1,6 @@
+if (process.env.NODE_ENV === 'development') {
+    require('./styles.scss');
+}
+
+require('./app/variation-toggler');
+require('./app/documentation-toggler');