+++ mote/views.py
@@ -1,12 +1,11 @@
-from cached_property import cached_property
-
-from django.views.generic.base import TemplateView
+from django.conf import settings
 from django.http import HttpResponse
 from django.template import TemplateDoesNotExist
 from django.template.loader import render_to_string
 from django.templatetags.static import PrefixNode
+from django.utils.functional import cached_property
 from django.utils.six.moves.urllib.parse import urljoin
-from django.conf import settings
+from django.views.generic.base import TemplateView
 
 from mote import PROJECT_PATHS
 from mote.models import Project, Aspect, Pattern, Element, Variation
@@ -115,7 +114,10 @@ class ElementIframeView(ElementBaseView):
     """Element view suitable for rendering in an iframe"""
 
     def get_template_names(self):
-        return ["mote/element/iframe.html"]
+        return [
+            "%s/mote/element/iframe.html" % self.element.project.id,
+            "mote/element/iframe.html"
+        ]
 
 
 class VariationBaseView(ElementBaseView):
@@ -142,4 +144,7 @@ class VariationIframeView(VariationBaseView):
     """Element view suitable for rendering in an iframe"""
 
     def get_template_names(self):
-        return ["mote/element/iframe.html"]
+        return [
+            "%s/mote/element/iframe.html" % self.element.project.id,
+            "mote/element/iframe.html"
+        ]