+++ mote/utils.py
@@ -18,3 +18,18 @@ def deepmerge(source, delta):
     """Return a deep merge of two dictionaries"""
 
     return _deepmerge(deepcopy(source), delta)
+
+
+# From http://stackoverflow.com/questions/5884066/hashing-a-python-dictionary
+def deephash(o):
+    if isinstance(o, (set, tuple, list)):
+	    return tuple([deephash(e) for e in o])
+
+    elif not isinstance(o, dict):
+        return hash(o)
+
+    new_o = deepcopy(o)
+    for k, v in new_o.items():
+        new_o[k] = deephash(v)
+
+    return hash(tuple(frozenset(sorted(new_o.items()))))