+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+next
+----
+#. Make it possible to override `iframe.html` for each project.
+
 0.1.1
 -----
 #. Inevitable rename to avoid name collision.