+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+next
+----
+#. Make it possible to override `iframe.html` for each project.
+#. New UI theme.
+#. Django template and RESTful API.
+
 0.1.1
 -----
 #. Inevitable rename to avoid name collision.