+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 next
 ----
 #. Make it possible to override `iframe.html` for each project.
+#. New UI theme.
+#. Django template and RESTful API.
 
 0.1.1
 -----